backend/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Backend for the LangGraph agent"
authors = [
    { name = "Philipp Schmid", email = "schmidphilipp1995@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph>=0.2.6",
    "langchain>=0.3.19",
    "python-dotenv>=1.0.1",
    "langgraph-sdk>=0.1.57",
    "langgraph-cli",
    "langgraph-cli[inmem]>=0.1.71",
    "langgraph-api",
    "fastapi",
    "google-genai",
    "nats-py>=2.10.0",
    "termcolor>=3.1.0",
    "langchain-ollama>=0.3.3",
    "langgraph-checkpoint>=2.1.0",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "transformers>=4.46.0",
    "torch>=2.1.0",
    "python-multipart>=0.0.6",
    "pandas>=2.3.0",
    "faiss-cpu",
    "chromadb",
    "ollama",
    "langchain-openai",
    "folium",
    "gradio",
    "langchain-google-vertexai",
    "langchain-neo4j",
    "neo4j"

]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
    "D100",
    "D101",
    "D104",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
